+++ Demo/Shared/Resources/Image+Demo.swift
@@ -23,6 +23,7 @@ extension Image {
     static var cornerRadius: Image { Image(systemName: "square") }
     static var cover: Image { Image(systemName: "rectangle.inset.fill") }
     static var data: Image { Image(systemName: "cylinder.split.1x2") }
+    static var detents: Image { Image(systemName: "rectangle.bottomthird.inset.fill") }
     static var dismiss: Image { Image(systemName: "xmark") }
     static var documentCamera: Image { Image(systemName: "doc.text.viewfinder") }
     static var download: Image { Image(systemName: "icloud.and.arrow.down") }
@@ -37,13 +38,13 @@ extension Image {
     static var optional: Image { Image(systemName: "questionmark.square.dashed") }
     static var palette: Image { Image(systemName: "paintpalette") }
     static var pickers: Image { Image(systemName: "list.bullet") }
-    static var sharing: Image { Image(systemName: "square.and.arrow.up") }
     static var pageControl: Image { Image(systemName: "ellipsis") }
     static var photo: Image { Image(systemName: "photo") }
     static var presentation: Image { Image(systemName: "square.stack") }
     static var qrCode: Image { Image(systemName: "qrcode") }
     static var shadow: Image { Image(systemName: "sun.dust.fill") }
-    static var sheet: Image { Image(systemName: "rectangle.bottomthird.inset.fill") }
+    static var sharing: Image { Image(systemName: "square.and.arrow.up") }
+    static var sheet: Image { Image(systemName: "rectangle.bottomhalf.inset.filled") }
     static var styles: Image { Image(systemName: "paintbrush.pointed") }
     static var swipeGesture: Image { Image(systemName: "hand.draw") }
     static var tapGesture: Image { Image(systemName: "hand.tap") }